Add files via upload
Mainly identifies the tutorial but a few things don't work. Collisions aren't fixed yet (working) and I need to add a couple of obstacles (people and spills?) note that pain mode is completely unusable for the time being. Try to figure out how to randomly open or close machines? (if you have time or are available.)